@@ -2,8 +2,9 @@ import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { environment } from '../../../environments/environment.development';
 import { Observable } from 'rxjs';
+import { Apollo, gql } from 'apollo-angular';
 
-const query = `
+const query = gql `
   query MyQuery {
     allPosts {
       id
@@ -13,6 +14,8 @@ const query = `
   }
 `
 
+
+
 @Component({
   selector: 'app-posts-table',
   standalone: false,
@@ -21,13 +24,21 @@ const query = `
 })
 export class PostsTableComponent implements OnInit {
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient, private apollo: Apollo) {}
 
   ngOnInit(): void {
-    this.getPosts().subscribe({
-      next: data => { console.log("NEXT", data); },
-      error: error => { console.log("ERR" + error); },
-    })
+    // this.getPosts().subscribe({
+    //   next: data => { console.log("NEXT", data); },
+    //   error: error => { console.log("ERR" + error); },
+    // });
+
+    this.apollo.watchQuery({
+      query: query
+    }).valueChanges.subscribe((data: any) => {
+      console.log("DATA", data.data);
+      console.log("ERR", data.error);
+      console.log("Loading", data.loading);
+    });
   }
 
   getPosts(): Observable<any> {
